Make splash screen trusted
Similarly to ag/12986397, touches were being blocked by
go/untrusted-touches when the user started interacting with the app
early on while the splash screen was showing.

Original intent of the window is to let touches pass through
(FLAG_NOT_TOUCHABLE) and it's controlled by the system, so we make it
trusted to allow touches to pass through.

Bug: 171772640
Bug: 158002302
Test: Launch app after clearing it from recents (to make sure there is
      no snapshot) and interact with it, verify no toast/logcat message
      about untrusted touches.
Change-Id: Ib20d8ce0975e547ec61bec1e3331e5d6aa6127bd

Former-commit-id: 29d8bbcf03bccef10ca83993c93255d4a2c38ebd