Revert "Enable building libandroid_runtime for Windows"
Building libandroid_runtime for Windows means building the
libnativehelper dependency for Windows as well. Since that library is
moving into the ART Mainline module it would mean having to provide it
as a prebuilt.

By removing this we avoid that, and the Windows variant isn't required
on the main branches, only in a downstream one.

This reverts commit b5c8033a2ebae6350c66581e22072c98a6afd21c [formerly 9e73bbbf4cbe406f67eb5029ccb8668fd691264f].

Test: `m` on git_master
Bug: 172649321
Change-Id: I5f0162e9b87529ee8b339bd24e5dee3e9bdd23a8

Former-commit-id: b0b9709abc1ce47de3fc73f4b52a19a90ed3502f