Init DownloadableSubscription with Confirmation Code
Currently carrier apps are not able to pass the confirmation code along
with the downloadable subscription because carrier apps do not have the
permission to set the CC. So carrier apps have to rely on LUI's scree
for entering the confirmation code. In order to make the process more
flexable for carriers, we decided to add a method to initialize a
Downloadable Subscription with the confirmation code.

Test: ag/13028582
Bug: 169276772
Change-Id: Icfa8b0f5b1699d24d8e39c3297ed25c7d1607739

Former-commit-id: 117f353a0b8c6faa2013fb8273e8e342f78cd54d