Excludes magnification switch button from system gesture area
When magnification switch button locates near to the system gesture
area, a user would feel hard to drag the button because the touch event
is dispatch to system gesture in priority.
To make the switch button easier to be dragged, we exclude the button
from system gesture area and adjust the layout size and the padding.

Bug: 172899084
Test: atest MagnificationModeSwitchTest
Change-Id: Ida99289ec63e6dc95d93eab979d9f8e985c3b889

Former-commit-id: 3a0b55820c8bbf4874885e0179ae073721f3da06