Run error callback when BiometricPrompt dismissed early
Currently, if BiometricPrompt is dismissed quickly (e.g. by tapping on
the top half of the screen to dismiss it while it's being shown) the
ongoing authentication session may not be canceled and the
onAuthenticationError callback may not be run. This commit fixes the
issue by ensuring that a dismissal reason is sent to callback if the
prompt is shown while in the "pending dismissal" state.

Test: atest com.android.systemui.biometrics
Test: Manual:
1. Using a custom test app, trigger BiometricPrompt to be shown
2. While the prompt is being shown, repeatedly tap top half of screen
3. Ensure the prompt is dismissed and onAuthenticationError is called

Bug: 173417042
Change-Id: I99284cb7fc413c0790f50056d553dc311117ce0c

Former-commit-id: a1cf4df73eeb05935655edae390c5015c6b6476a