Add BROADCAST_CLOSE_SYSTEM_DIALOGS to shell for tests
Tests that interact with the UI send Intent.ACSD before execution to
prevent the test from failing in case there are unexpected dialogs on
screen.

Those tests will be able to keep that via the instrumentation bridge
with the shell, given the shell has the proper permission to close
system dialogs.

Bug: 159105552
Test: Verify that shell has permission via adb shell dumpsys package
Change-Id: I07266573ae73d086446c6107248f5e187edcd64a

Former-commit-id: 4c9a6d5cdecc65c62579357240a207ab57c53065