Create single thread to play any Vibration
This thread takes in a CombinedVibrationEffect and plays all vibrations.

The Prebaked effect now also holds a fallback VibrationEffect, which is
resolved and scaled together with the original effect and is passed on
to the VibrateThread. All class attributes are not final, making this
effect immutable as all the others.

The Vibration now takes in a CombinedVibrationEffect, in preparation to
be used by the VibratorManagerService in multiple vibrators.

The new thread is replacing all vibration from VibratorService, which
means they are all triggering the HAL asynchronously.

Bug: 167946816
Bug: 131311651
Test: VibrationThreadTest, VibratorServiceTest, VibrationEffectTest, VibrationScalerTest
Change-Id: Ic27b35e63ca35ad47083f94da9ca7bd75b683d43

Former-commit-id: d6fae9a1e8ce2195ad26204daeb71222b87283e6