wifi: Add flag FLAG_IMMUTABLE in SAP notification
In S, it requires to specify explicitly either FLAG_MUTABLE
or FLAG_IMMUTABLE when creating a PendingIntent.

The SoftAp Notification is used to notify user the configuration changed.
And it supports to launch Setting page to show the current configuration.
Based on the flag description, the FLAG_MUTABLE should be considered when
the notification need to be used with inline reply or bubbles.
But it is not the use case on SoftAp Notification.
Use FLAG_IMMUTABLE.

Bug: 171719799
Test: Manuel Test
Change-Id: Id3df909f7754931e8018d7b8f6d07aba58158cbe

Former-commit-id: 6a334ed1cbb609de0b94cffadf21f5142ebc0594