Store mitigation call history before reboot
As part of the effort to better support failure mitigation across
reboots, track the mitigation calls across reboots by storing them
along with other salient parts of MonitoredPackage. These values
are relative to the current uptime of the system, so that they will
be accurate when the system uptime is reset at the next boot.

Also refactored code to allow for testing the reading and writing
of MonitoredPackage objects, and added tests for this.

Test: atest PackageWatchdogTest
Bug: 171951174
Change-Id: Ia96cf3892886d8d77193ffc278fa1eb584fecdd3

Former-commit-id: 25f5bf48bf5a7f2f99df0bd3ff7c9e01daffcdcd