Add a new method to notify system to detach the nav bar
For the attaching the nav bar to the app during quick switch, we
reparent the leash of the nav bar to the app the user swipes from and
restore it back when the quick switch gesture is finished.

There is a callback IRecentsAnimationController.finish() which seems to
be a good timing to restore the nav bar. However, it doesn't work well
in the live tile mode.

In the live tile mode, the overview won't call
IRecentsAnimationController.finish() until a new task is started and
this is why we need to add this API.

Bug: 139273001
Test: make
Change-Id: Iabea2fb222a2dacd57207b74d1778f93237268ea

Former-commit-id: 15900d8a3527b3a5e9cfd18a150634150c850e16