Suppress [CallbackInterface] metalava rule for mainline modules.
This rule is already suppressed for main framework code in
framework/Android.bp, but mainline uses a different set of metalava
flags.

Android API council says this is the right thing to do for libraries
which run only on Java 8+, which is true of all mainline modules and
framework code.

Bug: 172180369
Test: AppSearch BatchResultCallback now builds
Change-Id: If81cda012b4627c8018ae54a0f0932ee39d6fd96

Former-commit-id: c38a05e2bbc53ba1dc6efed9f343436087e7b09b