Unregister Touch Sensors for DOZE_AOD_DOCKED
DOZE_AOD_DOCKED should not have screen based trigger sensors
enabled. In fact, it can end up in inconsistent states currently.
When first entering it, the sensors are not enabled. But if the screen
then turns off (in a dark room) and then back on, the sensors are
left on from the prior state.

This change ensures that, when we ask to turn off touch based sensors,
they actually turn off.

Fixes: 172412931
Test: manual
Change-Id: I86f800cb75fee9f2d7e60655af4803e616a68fb2
Merged-In: I86f800cb75fee9f2d7e60655af4803e616a68fb2

Former-commit-id: 3995f0db93b6b5015fbdb6d89eefcc1952999b57