Update javadocs and remove unused hidden API constant
This change is part of the initiative to make managed account
provisioning support the admin-integrated flow.

We now deprecate ACTION_PROVISION_MANAGED_DEVICE as it will no
longer be used by admin apps. Instead, they will be directly
started by the managed account flow.

EXTRA_PROVISIONING_SKIP_USER_CONSENT is now deprecated as
it was only relevant for work profile on device owner which
is not a supported configuration.

EXTRA_PROVISIONING_KEEP_ACCOUNT_ON_MIGRATION and
EXTRA_PROVISIONING_LEAVE_ALL_SYSTEM_APPS_ENABLED are now also
allowed to be returned as extras to the intent result of the
GET_PROVISIONING_MODE admin app screen.

EXTRA_PROVISIONING_SKIP_USER_SETUP is now removed as it was
a hidden constant that was only relevant to split user mode,
which is no longer supported.

In addition, EXTRA_PROVISIONING_ACCOUNT_TO_MIGRATE documentation
is now updated to reflect that it can be also set by the
managed account provisioning initiator.

Test: compiled
Bug: 167955145
Change-Id: I8892d25209a9f1db17a1352390807372b956e704

Former-commit-id: 0d18e82854e3f1ba94149b35757cdabebeb1dd5c