Merge boundsChangeTransaction to syncTransaction when in blast sync
If the Window is currently in a blast sync and blast sync is enabled, we
need to ensure we don't lose the boundsChangeTransaction. We can't merge
it with the pending transaction and defer since it would cause the
transaction to never get applied and block.

Instead, just directly merge it with the sync transaction since the sync
transaction will contain the buffer. This will ensure it submits the
bounds change transaction at the same time as the buffer is applied

Test: Blast sync enabled with PIP that uses boundsChangeTransaction
Bug: 167202096
Change-Id: I585f361e83123fd1055bdaeeabbd203c9d1c83cf

Former-commit-id: b655e32207384f595af6f80aa6ebf4686ec896a2