Introduce Display `dontMoveToTop` property
If `dontMoveToTop` property is set to true, then Displays won't be
allowed to top when its root task moves to top.

Note: This CL fixes the IME/CarBoard bug with activities on cluster display.

Activities started on cluster display will move to the top of the task
activity stack and will break IME/CarBoard connectivity. This CL
introduces a new flag in order to disable activities started on cluster
display to move to the top, avoiding IME connection to break.

Manual test script:
1. Build and start emulator;
2. Start maps app;
3. Enter some text on top/right search text field and don't press ENTER
(just enter the text);
4. Start some activity on cluster display:
  `adb shell dumpsys car_service start-fixed-activity-mode 1 com.android.car com.android.car.am.ContinuousBlankActivity`
5. CarBoard should not disappear.

Bug: 169598740
Test: manual (see manual test script above)
Test: atest WmTests:TaskDisplayAreaTests
Test: atest DisplayContentTests
Test: atest DisplayWindowSettingsProviderTests

Change-Id: I7e5f067634a70df2093f20aa3077e0242fa817d5

Former-commit-id: ab616d9a71ecb76727813d41bd52fb76b7e9e4fa