Fix transcode_optimize flags
The optimize flag depended on transcode_enabled (sysprop only,
ignoring device_config) flag to take effect.

Now, we removed the 'optimize' flag and rely on a new non-persistent
sysprop that says if transcode was enabled on the current boot.

This means that a reboot is required for 'optimizations' to take
effect, right after enabling/disabling transcode. This is anyways a
better trade-off than ignoring the device_config transcode_enabled
flag entirely.

Test: Manual
Bug: 170488060
Change-Id: I77bbc30f9fae2cfc0085c32028859fdbbc9aaed7

Former-commit-id: 3952caa992edb7e67ea4753040476c1b5fd804cc