Add RemoteInput auth bypass flow
This change adds a way for a RemoteInputView activation to be allowed
even when the device is locked. If this is done, then authentication
will be deferred until the RemoteInput is sent. At that time, if
authentication is still necessary, the bouncer will be shown.

Authentication can be bypassed asynchronously while the RemoteInputView
is active via the AuthBypassPredicate passed to
NotificationRemoteInputManager#activateRemoteInput. If this predicate
returns true at the time that the RemoteInput is sent, then the send
will succeed, but the device will remain locked. If it returns false,
then the bouncer will be shown, at which point the user can choose to
unlock the device and send again, or dismiss the bouncer (and nothing
else will happen).

Test: manual
Bug: 170647553

Change-Id: Ie3b2089cefc6cfc2f5b6d218a61a51b1785ecf92

Former-commit-id: a7d7205e6ee1a5c0cf29fee9bcae0af34c368d49