add CpuFrequencies to BatterySaverPolicy
New CpuFrequencies methods were added to convert from string to map and
back to string. The Policy calss within BatterySaverPolicy is also
changed to hold the CpuFrequencies class directly instead of only
holding the system file map.

This change is in preperation for BatterySaverPolicy to freely convert
between Policy objects and BatterySaverPolicyConfig objects.

Bug: 172294448
Test: atest CpuFrequenciesTest
Test: atest PowerManagerTest
Test: atest BatterySaverPolicyTest
Test: atest CtsBatterySavingTestCases
Change-Id: Ib420404508c6839806a8e3b2beec59c90831a551

Former-commit-id: 8b317b15a3316cece158708db4913b05334c800d