Make window background blurs SystemApi
This CL makes windowBackgroundBlurEnabled, windowBackgroundBlurRadius
and WM.LP backgroundBlurRadius SystemApis to be used for bross-app blur.

WM.LP.backgroundBlurRadius is protected via a
Manifest.permission.USE_BACKGROUND_BLUR, which is checked both in the
system server and in SurfaceFlinger.

This CL also mokes WM.LP.FLAG_BLUR_BEHIND a SystemApi. It has been a
public API since 2009 and was deprected in 2011. It has been a no-op
ever since that. We are removing it from being a public api and will
use it as a system api in S

Bug: 167166562
Bug: 177524486
Test: manual - use PermissionController dialog to check SystemApis work
Change-Id: Icf98c3607b7d7ccdfb67b279e9fc5a83eba14204

Former-commit-id: e463ad6d1beb9ebe889ce4d8248ba1d3414e5dc1