Use rotated bounds as source touchable region for rotated app
The Layer in SF side will transform the given touchable region
according to the transformation matrix of surface. So if a window
is rotated, the raw touchable region should be the rotated bounds
instead of the bounds in display space. Then the raw region can
be rotated to the correct orientation. Otherwise the region is
like being rotated twice.

Bug: 175849666
Test: WindowStateTests#testUpdateInputWindowHandle
Change-Id: I266aaa21651c0bd5964e99d755382933d1336a9c

Former-commit-id: fa209a71789b01bc2b5da15f8614ad9903f64624