Expose role API for Settings on RoleManager.
Settings is a platform app so the role APIs for it wasn't system APIs,
but just @hide. Now that we are moving role into module, we need to
create system APIs for them.

RoleControllerManager is an implementation detail and may change in
soon when we move role logic into system server and leave only UI in
PermissionController, so we shouldn't expose it. Instead, we can
expose the new system APIs on RoleManager, and delegate to
RoleControllerManager internally.

Bug: 158736025
Test: presubmit
Change-Id: I6a3e5a26c3a133e4d1d8fc30a5d8508cd1788202

Former-commit-id: b6e4a3139e1f825b1ad2bcadb5f763927a2ec8b6