Remove MultiUserSwitch from KeyguardStatusBarView
Instead of MultiUserSwitch (a tappable view for opening the
KeyguardUserSwitcher), use an ImageView for showing the multi-user
avatar.

Also,

 - Remove KeyguardUserSwitcher from MultiUserSwitch. MultiUserSwitch
   will only be used for opening the QS user switcher now.

 - Cleanup other dependencies. For example,
   NotificationPanelViewController no longer needs QSDetailDisplayer.

NOTE: This change breaks KeyguardUserSwitcher since there is now no way
to open it. A follow-up change will introduce a new entry-point for
opening the switcher.

The QS user switcher will continue to work because it is handled by a
different instance of MultiUserSwitch (QSFooterView).

Test: setprop fw.max_users 1, on keyguard, confirm that user icon is not
      in status bar
Test: setprop fw.max_users 2, on keyguard, confirm that user icon is in
      status bar
Bug: 169783558
Change-Id: If93902985c3dc22be0cfa8b91d37cd9b3deced2f

Former-commit-id: 0c27903d9d1c245dce5cacbfbafd997f89d85297