Prototype tiles with side labels
This switches the tiles in QS to have the labels on the side, and
display in two columns.

Implemented:
* animation in expansion interleaves the tiles from QQS into QS, so the
first half of QQS moves to the left column
* Tapping on the label has the same effect as long pressing

Not implemented:
* animated shape between QQS and QS
* Edit screen

Landscape with media has a bug where scrolling pages resets to first
page. Also, animations in landscape with media don't look great.

To enable:
adb shell settings put secure sysui_side_labels 1
(0 to disable)

and restart SystemUI

Test: manual
Test: SystemUITests

Change-Id: I7468bc2c0b81e99a2de235a3b4312e678e968fd2

Former-commit-id: 703f295c2925ec38b1744eb5f9a0a3b3e972eec2