Add start / stop to location_time_zone_manager
Implement a path for reconfiguration of provider behavior at runtime.

Now the location_time_zone_manager can be stopped / started
from the command line with:

$ adb shell location_time_zone_manager stop
$ adb shell location_time_zone_manager start

This is useful during manual and automated (i.e. CTS) tests.

This change introduces LocationTimeZoneProviderController.destroy() and
all the associated plumbing / a new state needed to support "stop".

This commit removes NullLocationTimeZoneProvider and replaces it with a
LocationTimeZoneProviderProxy that does a similar thing.  It adds a
LocationTimeZoneProviderTest to replace the lost coverage.

Bug: 172934905
Bug: 152746105
Test: atest services/tests/servicestests/src/com/android/server/location/timezone/
Change-Id: I1ef7813ff2011251b2d8231a4e80f0374ee32c50

Former-commit-id: cd101fa5b972ba92e834197d6685fc652d32ee34