Mark android.net.sip code as deprecated.
This code has not been maintained in a long while and has a number of
pretty big flaws including: lack of IPV6 support, reliance on broadcasts
which are supported to wake up background receivers but don't.

Recommendation is for developers to use any number of other open source
SIP VOIP stacks available in the ecosystem.

It is also not currently supported on all Android devices.

Test: Make update-api; make
Bug: 175910634
Change-Id: I480d59ae4a09f74a03daa5b542fcfa1ed4a27925

Former-commit-id: 367d5e8e60d74ed8b2557245687bf7c525f129ee