Add GNSS capabilities and antenna broadcasts
Provide a system wide broadcast when GNSS capabilities change and when
GNSS antenna infos change so that clients have some way to detect this.

Antenna info was previously implemented as a listener, but after
evaluating the information, it is not protected, changes infrequently
with device configuration (if a foldable is unfolded for instance), and
clients need to see every change (not the previous model of only sending
changes while the client is foreground and GPS is enabled).

As a corollary, deprecate GNSS status events. These status messages
contain duplicate, inconsistent information, and do not scale for new
APIs. In addition, we don't have any clients that use these status
messages for any real purpose.

Bug: 153129152
Test: atest CtsLocationFineTestCases
Change-Id: Icc5bed8e37c7e8f168de57d2a019e94d96c0d3f2

Former-commit-id: 7a46eda83ee043d516467182bf49a3d55106e58f