Restrict StatusBarManager.collapsePanels() for targetSdk < S
In order to lock down the abuse vector (b/137274359) we need to restrict
this API for targetSdk < S too. So, using the same exemptions of
Intent.ACSD to allow legit use-cases, otherwise drop the call instead of
throwing because we're not gating it. Also allowing for holders of
permission STATUS_BAR since that's the permission we require when
targetSdk >= S.

Bug: 159105552
Test: CTS coming
Test: Verify call to collapsePanels() is not allowed
Test: Verify call to collapsePanels() while processing a trampoline with
      permission EXPAND_STATUS_BAR is allowed
Change-Id: I0b986f53df9b7773a33bdd2e57cdd241ff9026b0

Former-commit-id: 0a2b97d635d0ad6d61e99409b9817f8d4662ca71