Remove preserve surfaces when surface format changes.
When format is changed, there's no need to preserve surfaces. We can
just update the consumer to use the new format. There shouldn't be any
issues since the next dequeue will get a buffer with the new format. The
current buffer on screen won't change and the layer properties aren't changed
when format is updated. There's no need to recreate a new surface.

Instead, move the format updating logic to the client in VRI and send
the new format to BLASTBufferQueue so it can update the consumer.

Test: go/wm-smoke
Bug: 177557720
Change-Id: I26806cfd4b482e8b6165045f4af3c55a5955da0c

Former-commit-id: 41117a4be71982b137ebc713b1ddc8a0b2df955f