UserSystemPackageInstaller auto-treats static overlays
The UserSystemPackageInstaller installs packages on new users
based on their configuration in an allowlist. An exception was
made for auto-generated RROs, which were instead installed based
on their overlay target package's configuration (thus avoiding
the need for manually allowlisting each such overlay).

This is now expanded to all static overlays (which should
include auto-generated RROs, according to the bug). That is:
Static overlays no longer need to be mentioned in the list,
but will instead be installed based on the listing for their
overlay target package.

Test: atest UserSystemPackageInstallerTest
Fixes: 172956245
Change-Id: I6b974bba20ee059a8f744092db7c3441580d327f

Former-commit-id: 1f04060be06545c86c0d377b8dd81e6f4cbb1c9f