Mark RotationLayer as trusted for untrusted touches
Its content is generated by the system (a snapshot of the screen IIUC,
please correct me if I'm wrong!) and the layer was originally
pass-through.

Touches are being blocked in some tests, which is causing breakages, so
reverting to the previous behavior for this layer, which is to let
touches pass through.

Bug: 176926757
Bug: 174269206
Bug: 158002302
Test: Rotate device, tap the screen frenetically during animation,
      observer no log from touches being blocked.
Change-Id: Iea970e6a4793dd991e2a666c04bb508d7fa240ed

Former-commit-id: 032de31170be7915541e509502265323f8891e0d