Convert DomainVerificationService.Connection to synchronous set
Unfortunately it's possible to call this during PackageManagerService
initialization, so this needs to be attached before PMS gets created.

Rather than using a deferred injection, this makes PMS call
setConnection to directly set the callback. A better isolated might be
worth exploring in the future, but overall it shouldn't really matter.

Exempt-From-Owner-Approval: Already approved by owners on main branch

Bug: 163565712

Test: com.android.server.pm.test.domain.verify

Change-Id: Ib60f5560ecaa510474228aa1c38eff5d6041ab58

Former-commit-id: 47dbabfc14420ff54646df05039657df43901554