Deprecate RoleControllerService related system APIs.
The RoleControllerService is an internal implementation detail of
role, and was exposed as system API in platform before modularization
and the ability to have updatable code in system server.

Now that we have our updatable system server JAR, this should no
longer be exposed and we may refactor the implementation for better
performance/stability later. Hence deprecate them now in S for the
earliest possible removal of these APIs.

Bug: 158736025
Test: presubmit
Change-Id: I1358e16ad0e9cf875316d93dd2152001c13b84dd

Former-commit-id: 61545a7efd635fb0701c4376e5abf6611e76cd4f