Avoid detecting WebView as TYPE_RECYCLING
Currently ScrollCaptureIntrnal only checks for >1 child views
but not >0 child views. Because of this, WebView is considered
a RecyclerView, but this will not work without any Views to
track.

This change classifies these as "OPAQUE" so they can be handled
specially. Until support for these is implemented they will be
skipped as targets.

Bug: 179253269
Test: Open Chrome to a scrollable page, take screenshot
Change-Id: I0b00598031879d2fa7c1295440b3de1e0e453031

Former-commit-id: d84643adcc2dbbfdbe0ca92c9000b226bdbd7210