Invoke organizer events when a display is added / removed.
Before, when a display is added, we don't attach existing organizer to
it. Now, we checked if there is an existing one by checking the feature
id.

At ag/13237565, we tried to do it at onParentChanged(), but
mDisplayContent is not set for DAs when DAs are first attached. Change
to do it from DisplayContent when it is ready.

Fix: 179102157
Test: atest WmTests:DisplayAreaTest
Change-Id: Ic2bb7447e9a1408ccdc17f79271b2cf58b2878c1

Former-commit-id: 2c133360a4797c1c13f7ee71841960ac9c728703